remove machine scope from exportMarkdownPreviewStyle config
Even though there is exactly one place on machine where the preview style is defined, it's okay to be able to set this on workspace basis. The extension will generate a new theme based on workspace config when the workspace is opened.

Worth noting that when multiple workspaces with different theme configs then only the one opened later will be applied since there can only be one theme file per machine.